TPGSWE-21206: Fix handling of SerDes app restart in PHY driver
Upon detecting a SerDes app restart, both the PLL and SerDes lanes
were being reconfigured. However, reconfiguring the PLL at this stage
is problematic and may disrupt traffic transmission and reception.
As a temporary WA, the PLL reconfiguration step has been removed.

This commit also includes triggering the PHY state machine to
immediately update the link status when a link-down event is forced,
as well as preventing the kernel from sending packets when the
carrier is off.