TPGSWE-21476: Keep PCS TX and switch port enabled when link is down
Disabling the switch port when the link goes down can make it non-functional
if data is still in the switch buffers. To avoid this, both the switch port
and PCS TX remain enabled during link-down events. This allows the switch
buffers to flush and prevents backpressure that could affect traffic on the
secondary port.

Additionally, to avoid skewing MAC TX and switch TX statistics due to
broadcast/multicast traffic, packet flooding is disabled for the affected
port when the link is down.