MAINT: Move kalloc out of spinlock
GFP_KERNEL memory allocations should not be within spin locks as it may sleep.
Instead, we encapsulate only the part of the function that modifies the struct
member needing protection.

The RCU calls don't strictly need the spin lock, but it's included to avoid
differences between the counter and hash table.

Signed-off-by: Kim Holdt <Kim.Holdt@analog.com>